+++ skysail.api/src/io/skysail/api/text/TranslationStore.java
@@ -1,11 +1,11 @@
 package io.skysail.api.text;
 
-import java.util.*;
+import java.util.Locale;
+import java.util.Optional;
 
+import org.osgi.annotation.versioning.ProviderType;
 import org.osgi.framework.BundleContext;
 import org.restlet.Request;
-
-import aQute.bnd.annotation.ProviderType;
 
 /**
  * A translation store implements a way to retrieve (and persist or update) translations
@@ -22,7 +22,7 @@ public interface TranslationStore {
     Optional<String> get(String key, ClassLoader cl, Request request);
 
     Optional<String> get(String key, ClassLoader cl, Request request, Locale locale);
-    
+
     boolean persist(String key, String message, Locale locale, BundleContext bundleContext);
 
 }