+++ skysail.domain/src/io/skysail/domain/core/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * skysail core domain: Models of applications, entities, fields, etc.
+ */
+@Version("3.0.0")
+package io.skysail.domain.core;
+
+import org.osgi.annotation.versioning.Version;