+++ skysail.server.db/src/io/skysail/server/db/DbService.java
@@ -3,10 +3,11 @@ package io.skysail.server.db;
 import java.util.List;
 import java.util.Map;
 
+import org.osgi.annotation.versioning.ProviderType;
+
 import com.orientechnologies.orient.core.metadata.schema.OType;
 import com.orientechnologies.orient.core.record.impl.ODocument;
 
-import aQute.bnd.annotation.ProviderType;
 import io.skysail.domain.Identifiable;
 import io.skysail.domain.core.ApplicationModel;