+++ skysail.server.db/src/io/skysail/server/db/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * db layer definition.
+ */
+@Version("8.0.0")
+package io.skysail.server.db;
+
+import org.osgi.annotation.versioning.Version;