+++ skysail.server.db/src/io/skysail/server/db/versions/impl/VersionsRepository.java
@@ -1,14 +1,18 @@
 package io.skysail.server.db.versions.impl;
 
-import io.skysail.domain.core.repos.DbRepository;
-import io.skysail.server.db.*;
-
 import java.util.Collections;
 
-import lombok.extern.slf4j.Slf4j;
-import aQute.bnd.annotation.component.*;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
-@Component(immediate = true, properties = "name=VersionsRepository")
+import io.skysail.domain.core.repos.DbRepository;
+import io.skysail.server.db.DbClassName;
+import io.skysail.server.db.DbService;
+import io.skysail.server.db.GraphDbRepository;
+import lombok.extern.slf4j.Slf4j;
+
+@Component(immediate = true, property = "name=VersionsRepository")
 @Slf4j
 public class VersionsRepository extends GraphDbRepository<ComponentDbVersion> implements DbRepository {