+++ skysail.server.db/src/io/skysail/server/db/versions/package-info.java
@@ -1,4 +1,7 @@
+/**
+ * db versions
+ */
 @Version("1.0.0")
-package io.skysail.server.um.shiro.app;
+package io.skysail.server.db.versions;
 
 import org.osgi.annotation.versioning.Version;