+++ skysail.server.um.httpbasic/src/io/skysail/server/um/httpbasic/app/package-info.java
@@ -1,2 +1,4 @@
-@aQute.bnd.annotation.Version("0.1.0")
+@Version("0.1.0")
 package io.skysail.server.um.httpbasic.app;
+
+import org.osgi.annotation.versioning.Version;