+++ skysail.server/src/io/skysail/server/resources/package-info.java
@@ -1,5 +1,7 @@
 /**
  * Package holding standard resources.
  */
-@aQute.bnd.annotation.Version("0.1.0")
+@Version("0.1.0")
 package io.skysail.server.resources;
+
+import org.osgi.annotation.versioning.Version;