+++ skysail.api/etc/changelog.md
@@ -4,6 +4,7 @@ new in
 ------
 
  * added getLoginPath() and getLogoutPath() methods to Authentication Service
+ * removed biz.aQute.bnd.annotations
 
 2.1.0
 ------